Library for G+ calls
G+ methods are now in a library so that they can be used by other
packages.
